+++ src/database/models/roles/capability.rs
@@ -77,4 +77,9 @@ impl Capability {
 }
 
 /// All the capabilities of the application
-pub const CAPABILITIES: [&str; 3] = ["role:manage", "post:create", "user:manage"];
+pub const CAPABILITIES: [&str; 4] = [
+    "role:manage",
+    "post:create",
+    "user:manage",
+    "user:manage_role",
+];